requirements.txt
asgiref==3.4.1
boto3==1.20.19
botocore==1.23.19
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
jmespath==0.10.0
Markdown==3.3.4
Pillow==8.3.1
psycopg2-binary==2.9.2
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.3.0
